Modified the "QwCopyArray" function in QwUtil.h to use a "CopyFrom" member function of hardware channels.
The previous use of std::copy in "QwCopyArary" failed to copy the configuration information for the C-style arrays of hardware channels in the BPM and QPD classes.  I think that is because the assignment operator is used to only copy the event-specific data, to allow mathematic operations between different hardware channels.  The result is that the internal data elements of the BPM and QPD classes were not set properly in copies, and we could not calculate BPM/QPD yields, asymmetries, or correlations.

I added a "CopyFrom(value)" member function to VQwDataElement, VQwHardwareChannel, and the scaler and VQWK channel classes; this copies the channel configuration information from the value object into the current object.  It is based on the copu constructors for the respective classes.

I then rewrote the QwCopyArray function to use "CopyFrom".  In the process, I realized it would be easiest to change the C-style arrays used for the internal data members to be a std::array.  This has been done for QwBPMCavity,QwBPMStripline, QwCombinedBPM, and QwQPD.