Fixed a typo in the "qwparity" instructions in README.md and found a bug for ALMA9 compilation.
In the README.md, the instructions for running "qwparity" had a typo;
the commandline flag to specify the data handler map file should be
"--datahandlers".

Then, running with the data handlers active on the ALMA9 machines, we
found a segfault during the program initiallization due to a
non-initialized pointer in the VQwDataHandler class.  That pointer is
now explicitly zeroed during construction.

The correlator and the linear regression corrector are now executed, but
we need to verify that their opperation is correct.